requirements.txt
neo4j==5.28.2 
pinecone-client==2.2.0
pyvis==0.3.1
networkx==3.1
tqdm==4.65.0
httpx==0.27.0
sentence-transformers 
